pyproject.toml
[project]
name = "de-test-etl"
version = "0.1.0"
description = "Demo ETL pipeline project (test task)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["etl", "db"]

[tool.hatch.build.targets.wheel]
# include non-Python files in the wheel
include = [
    "alembic.ini",
    "data/orders_data.json",
]
